Break apart bets file
